requirements.txt
Django==4.0.3
celery==5.2.6
django-filter==21.1
APScheduler==3.9.1
django-allauth==0.50.0
python-dotenv==0.20.0